Create ms_rewards_vpn_dict.json
Use this file to run with the  vpn. 

Use api.nodrvpn.com/server for json list of all servers and names. Use that to copy and paste what servers you wish to use.
Position corresponds to the account number as displayed in the other json login file. Whatever order you have entered in that is the order that the vpn will login to the account.